+++ nbproject/private/private.xml
@@ -2,11 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/StarmanW/Documents/NetBeansProjects/FioreFlowershop/src/CustomizedFloralArrangement/CustomizedFloralArrangementTestMain.java</file>
-            <file>file:/C:/Users/StarmanW/Documents/NetBeansProjects/FioreFlowershop/src/CustomizedFloralArrangement/PickUpQueue.java</file>
-            <file>file:/C:/Users/StarmanW/Documents/NetBeansProjects/FioreFlowershop/src/CustomizedFloralArrangement/PickUpQueueInterface.java</file>
-            <file>file:/C:/Users/StarmanW/Documents/NetBeansProjects/FioreFlowershop/src/CustomizedFloralArrangement/Utility.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
